+++ dafni_cli/model/model.py
@@ -1,29 +1,18 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from datetime import date, datetime
-from typing import Any, ClassVar, List, Optional
+from typing import ClassVar, List, Optional
 
 import click
 
+from dafni_cli.api.auth import Auth
 from dafni_cli.api.parser import ParserBaseObject, ParserParam, parse_datetime
 from dafni_cli.consts import (
     CONSOLE_WIDTH,
-    INPUT_DEFAULT_HEADER,
-    INPUT_DESCRIPTION_HEADER,
-    INPUT_DESCRIPTION_LINE_WIDTH,
-    INPUT_MAX_HEADER,
-    INPUT_MIN_HEADER,
-    INPUT_MIN_MAX_COLUMN_WIDTH,
-    INPUT_TITLE_HEADER,
-    INPUT_TYPE_COLUMN_WIDTH,
-    INPUT_TYPE_HEADER,
-    OUTPUT_FORMAT_COLUMN_WIDTH,
-    OUTPUT_FORMAT_HEADER,
-    OUTPUT_NAME_HEADER,
-    OUTPUT_SUMMARY_COLUMN_WIDTH,
-    OUTPUT_SUMMARY_HEADER,
     TAB_SPACE,
 )
-from dafni_cli.utils import optional_column_new, prose_print
+from dafni_cli.model.inputs import ModelInputs
+from dafni_cli.model.outputs import ModelOutputs
+from dafni_cli.utils import prose_print
 
 
 @dataclass
@@ -68,307 +57,6 @@ class ModelMetadata(ParserBaseObject):
         ParserParam("publisher", "publisher", str),
         ParserParam("source_code", "source_code", str),
     ]
-
-
-@dataclass
-class ModelAuth(ParserBaseObject):
-    """Dataclass representing the access the user has for a DAFNI model
-
-    Attributes:
-        view (bool): View access
-        read (bool): Read access
-        update (bool): Update access
-        destroy (bool): Deletion access
-        reason (str): Reason user has access to view this model
-
-
-        The following are only present for the /model/<version_id> endpoint
-        (but are guaranteed for it)
-        --------
-        asset_id (Optional[str]): ID of the model
-
-
-        The following are only present for the /models endpoint
-        (but are guaranteed for it)
-        --------
-        role_id (Optional[str]): Role ID of the user
-        name (Optional[str]): Name associated with the auth type
-    """
-
-    view: bool
-    read: bool
-    update: bool
-    destroy: bool
-    reason: str
-
-    asset_id: Optional[str] = None
-    name: Optional[str] = None
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("view", "view"),
-        ParserParam("read", "read"),
-        ParserParam("update", "update"),
-        ParserParam("destroy", "destroy"),
-        ParserParam("reason", "reason", str),
-        ParserParam("asset_id", "asset_id", str),
-    ]
-
-
-@dataclass
-class ModelDataslot(ParserBaseObject):
-    """Dataclass representing an input dataslot for a DAFNI
-       model
-
-    Attributes:
-        name (str): Name of the slot
-        path (str): Path of the file expected in the container
-        defaults (List[str]): List of default dataset ids
-        required (str): Whether the slot is required
-        description (Optional[str]): Description of the slot if applicable
-    """
-
-    name: str
-    path: str
-    defaults: List[str]
-    required: bool
-    description: Optional[str] = None
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("name", "name", str),
-        ParserParam("path", "path", str),
-        ParserParam("defaults", "default"),
-        ParserParam("required", "required"),
-        ParserParam("description", "description", str),
-    ]
-
-
-@dataclass
-class ModelParameter(ParserBaseObject):
-    """Dataclass representing an input parameter for a DAFNI model
-
-    Attributes:
-        name (str): Name of the parameter
-        type (str): Type of the parameter
-        title(str): Title of the parameter
-        required (bool): Whether the parameter is required
-        description (str): Description of the model parameter
-        default (Any): Default value of the parameter
-    """
-
-    name: str
-    type: str
-    title: str
-    required: bool
-    description: str
-    default: Optional[Any] = None
-
-    # TODO: Left over from refactor, may no longer be needed
-    # (printed in format_parameters in ModelInputs but untouched for 2 years)
-    min: Optional[str] = None
-    max: Optional[str] = None
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("name", "name", str),
-        ParserParam("type", "type", str),
-        ParserParam("title", "title"),
-        ParserParam("required", "required"),
-        ParserParam("description", "description"),
-        ParserParam("default", "default"),
-        ParserParam("min", "min"),
-        ParserParam("max", "max"),
-    ]
-
-
-@dataclass
-class ModelInputs(ParserBaseObject):
-    """Dataclass representing inputs for a DAFNI model
-
-    Attributes:
-        parameters (List[ModelParameters]): List of parameters for the model
-        dataslots (List[ModelDataslot]): List of dataslots for the model
-    """
-
-    parameters: List[ModelParameter]
-    dataslots: List[ModelDataslot] = field(default_factory=list)
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("parameters", "parameters", ModelParameter),
-        ParserParam("dataslots", "dataslots", ModelDataslot),
-    ]
-
-    # TODO: Couldn't we use tabulate or something for this?
-    @staticmethod
-    def params_table_header(title_column_width: int, default_column_width: int) -> str:
-        """Formats the header row and the dashed line for an input parameters
-           table
-
-        Args:
-            title_column_width (int): Width of the title column to fit the
-                                      longest title (or "title") plus a buffer
-            default_column_width (int): Width of the default column to fit the
-                                        longest default (or "default") plus a
-                                        buffer
-
-        Returns:
-            str: Formatted string for the header and dashed line of the
-                 parameters table
-        """
-
-        header = (
-            f"{INPUT_TITLE_HEADER:{title_column_width}}"
-            f"{INPUT_TYPE_HEADER:{INPUT_TYPE_COLUMN_WIDTH}}"
-            f"{INPUT_MIN_HEADER:{INPUT_MIN_MAX_COLUMN_WIDTH}}"
-            f"{INPUT_MAX_HEADER:{INPUT_MIN_MAX_COLUMN_WIDTH}}"
-            f"{INPUT_DEFAULT_HEADER:{default_column_width}}"
-            f"{INPUT_DESCRIPTION_HEADER}\n"
-            + "-"
-            * (
-                title_column_width
-                + INPUT_TYPE_COLUMN_WIDTH
-                + 2 * INPUT_MIN_MAX_COLUMN_WIDTH
-                + default_column_width
-                + INPUT_DESCRIPTION_LINE_WIDTH
-            )
-            + "\n"
-        )
-        return header
-
-    def format_parameters(self) -> str:
-        """Formats input parameters for a model into a string which prints as
-           a table
-
-        Returns:
-            str: Formatted string that will appear as a table when printed
-        """
-
-        titles = [parameter.title for parameter in self.parameters] + ["title"]
-        defaults = ["default"]
-        for parameter in self.parameters:
-            if parameter.default is not None:
-                defaults.append(str(parameter.default))
-        title_column_width = len(max(titles, key=len)) + 2
-        default_column_width = len(max(defaults, key=len)) + 2
-        # Setup headers
-        params_table = ModelInputs.params_table_header(
-            title_column_width, default_column_width
-        )
-        # Populate table
-        for parameter in self.parameters:
-            params_table += f"{parameter.title:{title_column_width}}{parameter.type:{INPUT_TYPE_COLUMN_WIDTH}}"
-            params_table += optional_column_new(
-                parameter.min, INPUT_MIN_MAX_COLUMN_WIDTH
-            )
-            params_table += optional_column_new(
-                parameter.max, INPUT_MIN_MAX_COLUMN_WIDTH
-            )
-            params_table += optional_column_new(parameter.default, default_column_width)
-            params_table += f"{parameter.description}\n"
-        return params_table
-
-    def format_dataslots(self) -> Optional[str]:
-        """Formats input data slots to print in a clear way
-
-        Returns:
-            Optional[str]: Formatted string that will present the dataslots
-                           clearly when printed
-        """
-
-        if self.dataslots:
-            dataslots_list = ""
-            for dataslot in self.dataslots:
-                dataslots_list += "Name: " + dataslot.name + "\n"
-                dataslots_list += "Path in container: " + dataslot.path + "\n"
-                dataslots_list += f"Required: {dataslot.required}\n"
-                dataslots_list += "Default Datasets: \n"
-                for default in dataslot.defaults:
-                    # TODO print name using API call to databases
-                    dataslots_list += "ID: " + default + TAB_SPACE
-                    # dataslots_list += f'ID: {default["uid"]}' + TAB_SPACE
-                    # dataslots_list += f'Version ID: {default["versionUid"]}' + TAB_SPACE
-                dataslots_list += "\n"
-            return dataslots_list
-        return None
-
-
-@dataclass
-class ModelOutputDataset(ParserBaseObject):
-    """Dataclass containing information on an output dataset from a model
-
-    Attributes:
-        name (str): Name of the dataset output
-        type (str): Type of the dataset output
-        description (str): Description of the dataset output
-    """
-
-    name: Optional[str] = None
-    type: Optional[str] = None
-    description: Optional[str] = None
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("name", "name", str),
-        ParserParam("type", "type", str),
-        ParserParam("description", "description", str),
-    ]
-
-
-@dataclass
-class ModelOutputs(ParserBaseObject):
-    """Dataclass representing outputs for a DAFNI model
-
-    Attributes:
-        datasets (List[ModelOutputDataset]): List of output datasets
-    """
-
-    datasets: List[ModelOutputDataset]
-
-    _parser_params: ClassVar[List[ParserParam]] = [
-        ParserParam("datasets", "datasets", ModelOutputDataset),
-    ]
-
-    @staticmethod
-    def outputs_table_header(name_column_width: int) -> str:
-        """Formats the header row and the dashed line for the output parameters
-           table
-
-        Args:
-            name_column_width (int): Width of the name column to fit the
-                                     longest name (or "name") plus a buffer
-
-        Returns:
-            str: Formatted string for the header and dashed line of the outputs table
-        """
-        header = (
-            f"{OUTPUT_NAME_HEADER:{name_column_width}}"
-            f"{OUTPUT_FORMAT_HEADER:{OUTPUT_FORMAT_COLUMN_WIDTH}}"
-            f"{OUTPUT_SUMMARY_HEADER}\n"
-            + "-"
-            * (
-                name_column_width
-                + OUTPUT_FORMAT_COLUMN_WIDTH
-                + OUTPUT_SUMMARY_COLUMN_WIDTH
-            )
-            + "\n"
-        )
-        return header
-
-    def format_outputs(self) -> str:
-        """Formats output files into a string which prints as a table
-
-        Returns:
-            str: Formatted string that will appear as a table when printed
-        """
-        names = [dataset.name for dataset in self.datasets] + ["Name"]
-        max_name_length = len(max(names, key=len)) + 2
-        outputs_table = ModelOutputs.outputs_table_header(max_name_length)
-
-        # The dataset outputs fields are not mandatory and any or all of them might not
-        # exist. Unset fields will be reported as "Unknown" in the formatted output.
-        for dataset in self.datasets:
-            outputs_table += f"{dataset.name or 'Unknown':{max_name_length}}"
-            outputs_table += f"{dataset.type or 'Unknown':{OUTPUT_FORMAT_COLUMN_WIDTH}}"
-            outputs_table += optional_column_new(dataset.description)
-            outputs_table += "\n"
-        return outputs_table
 
 
 @dataclass
@@ -440,8 +128,8 @@ class Model(ParserBaseObject):
                                   model version
         version_history (List[ModelVersion]): Full version history of the
                                               model
-        auth (ModelAuth): Authentication credentials giving the permissions
-                          the current user has on the model
+        auth (Auth): Authentication credentials giving the permissions the
+                     current user has on the model
         ingest_completed_date (datetime or None): Date and time the model
                                            finished ingesting if applicable
         metadata (ModelMetadata): Metadata of the model
@@ -469,7 +157,7 @@ class Model(ParserBaseObject):
     version_message: str
     version_tags: List[str]
     version_history: List[ModelVersion]
-    auth: ModelAuth
+    auth: Auth
     ingest_completed_date: Optional[datetime] = None
 
     api_version: Optional[str] = None
@@ -501,7 +189,7 @@ class Model(ParserBaseObject):
         ParserParam("version_message", "version_message", str),
         ParserParam("version_tags", "version_tags"),
         ParserParam("version_history", "version_history", ModelVersion),
-        ParserParam("auth", "auth", ModelAuth),
+        ParserParam("auth", "auth", Auth),
         ParserParam("ingest_completed_date", "ingest_completed_date", parse_datetime),
         ParserParam("_metadata", "metadata", ModelMetadata),
         ParserParam("api_version", "api_version", str),
@@ -644,3 +332,17 @@ class Model(ParserBaseObject):
             click.echo(f"Version message: {version.version_message}")
             click.echo(f"Version tags: {', '.join(version.version_tags)}")
             click.echo("")
+
+
+# The following methods mostly exists to get round current python limitations
+# with typing (see https://stackoverflow.com/questions/33533148/how-do-i-type-hint-a-method-with-the-type-of-the-enclosing-class)
+def parse_models(model_dictionary_list: List[dict]) -> List[Model]:
+    """Parses the output of get_all_models and returns a list of Model
+    instances"""
+    return ParserBaseObject.parse_from_dict_list(Model, model_dictionary_list)
+
+
+def parse_model(model_dictionary: dict) -> Model:
+    """Parses the output of get_model and returns a list of Model
+    instances"""
+    return ParserBaseObject.parse_from_dict(Model, model_dictionary)
